+++ resolwe_bio/tests/processes/test_clustering.py
@@ -19,19 +19,59 @@ class ClusteringProcessorTestCase(BioProcessTestCase):
         self.run_process('clustering-hierarchical-genes-etc', inputs)
 
     def test_hc_clustering_samples(self):
-        inputs = {'exp': 'hs_expressions_1.tab.gz', 'exp_type': 'Log2', 'exp_name': 'Expression', 'source': 'UCSC'}
+        inputs = {'exp': 'clustering_expressions_1.tab.gz',
+                  'exp_type': 'Log2',
+                  'exp_name': 'Expression',
+                  'source': 'UCSC'}
         expression_1 = self.run_process('upload-expression', inputs)
 
-        inputs = {'exp': 'hs_expressions_2.tab.gz', 'exp_type': 'Log2', 'exp_name': 'Expression', 'source': 'UCSC'}
+        inputs = {'exp': 'clustering_expressions_2.tab.gz',
+                  'exp_type': 'Log2',
+                  'exp_name': 'Expression',
+                  'source': 'UCSC'}
         expression_2 = self.run_process('upload-expression', inputs)
 
-        inputs = {'exp': 'hs_expressions_3.tab.gz', 'exp_type': 'Log2', 'exp_name': 'Expression', 'source': 'UCSC'}
+        inputs = {'exp': 'clustering_expressions_3.tab.gz',
+                  'exp_type': 'Log2',
+                  'exp_name': 'Expression',
+                  'source': 'UCSC'}
         expression_3 = self.run_process('upload-expression', inputs)
 
-        inputs = {'exps': [expression_1.pk, expression_2.pk, expression_3.pk]}
+        inputs = {'exps': [expression_1.pk, expression_2.pk, expression_3.pk],
+                  'genes': [' A1BG E2f4 A2ML1 A2MP1 A3GALT2 A4GALT AADACL4 AADACL3 AADACL2 AADAC'],
+                  'genes_source': 'UCSC'}
+
         clustering = self.run_process('clustering-hierarchical-samples', inputs)
 
-        saved_json, test_json = self.get_json('sample_custer_data.json.gz', clustering.output['cluster'])
+        saved_json, test_json = self.get_json('sample_cluster_data.json.gz', clustering.output['cluster'])
         self.assertEqual(test_json['linkage'], saved_json['linkage'])
-        self.assertTrue('samples_names' in test_json)
+        self.assertTrue('order' in test_json)
+
+    def test_hc_clustering_genes(self):
+        inputs = {'exp': 'clustering_expressions_1.tab.gz',
+                  'exp_type': 'Log2',
+                  'exp_name': 'Expression',
+                  'source': 'UCSC'}
+        expression_1 = self.run_process('upload-expression', inputs)
+
+        inputs = {'exp': 'clustering_expressions_2.tab.gz',
+                  'exp_type': 'Log2',
+                  'exp_name': 'Expression',
+                  'source': 'UCSC'}
+        expression_2 = self.run_process('upload-expression', inputs)
+
+        inputs = {'exp': 'clustering_expressions_3.tab.gz',
+                  'exp_type': 'Log2',
+                  'exp_name': 'Expression',
+                  'source': 'UCSC'}
+        expression_3 = self.run_process('upload-expression', inputs)
+
+        inputs = {'exps': [expression_1.pk, expression_2.pk, expression_3.pk],
+                  'genes': [' A1BG E2f4 A2ML1 A2MP1 A3GALT2 A4GALT AADACL4 AADACL3 AADACL2 AADAC'],
+                  'genes_source': 'UCSC'}
+
+        clustering = self.run_process('clustering-hierarchical-genes', inputs)
+
+        saved_json, test_json = self.get_json('gene_cluster_data.json.gz', clustering.output['cluster'])
+        self.assertEqual(test_json['linkage'], saved_json['linkage'])
         self.assertTrue('order' in test_json)